refactor: moved words to text_generator and added text_generator module
